Deduplicate JS code for benchmarks 4 and 5.
