+++ src/tests_webassembly/resources/buyer/buyer-chromium.template.js
@@ -1,0 +1,1 @@
+{% extends "buyer-common.template.js" %}
