--- src/tests_webassembly/resources/buyer/cxx-src/to_ints.py
@@ -1,8 +1,0 @@
-import sys
-
-sys.stdout.write('[')
-for c in sys.stdin.buffer.read():
-    sys.stdout.write(str(c))
-    sys.stdout.write(',')
-sys.stdout.write(']')
-
