+++ src/tests_webassembly/resources/buyer/index.html
@@ -21,7 +21,7 @@ if (!navigator.joinAdInterestGroup) {
   navigator.joinAdInterestGroup({
     owner: "https://fledge-tests.creativecdn.net:9021",
     name: "tc-ig",
-    biddingLogicUrl: "https://fledge-tests.creativecdn.net:9021/buyer.js",
+    biddingLogicUrl: "https://fledge-tests.creativecdn.net:9021/buyer-chromium.js",
     userBiddingSignals: {key: "user bidding signals "},
     ads: [
       {
